requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
bitsandbytes==0.44.1
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
dacite==1.8.1
DAWG-Python==0.7.2
debugpy==1.8.6
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
ImageHash==4.3.1
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multimethod==1.12
murmurhash==1.0.10
narwhals==1.9.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pandas-profiling==3.6.6
parso==0.8.4
patsy==0.5.6
peft==0.13.0
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc
ruff==0.6.8
safetensors==0.4.5
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.6
statsmodels==0.14.4
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
thinc==8.3.2
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
visions==0.7.6
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
widgetsnbextension==4.0.13
wordcloud==1.9.3
wrapt==1.16.0
ydata-profiling==4.10.0
